Remove rendered div from Calendar component
The calendar component now no longer renders a div. This way the
component could be used in any React render environment such as React
native. It also does not have to forward a bunch of DOM props to the div
anymore. The user can instead just render the div themselves:

Previously:
<Calendar className="bg-white">{/* ... */}</Calendar>
Now:
<Calendar><div className="bg-white">{/* ... */See: #1}</div></Calendar>

See: #1