Change signature of setDate
Set date can now be called with a date directly or with an update
function that takes the old date and returns a new date, similar to
react state update functions (in fact it is a React setState
underneath). This makes it easier to implement going forward for a
specified amount of days. This also works very well with the date-fns/fp
package.